requirements.txt
cat > requirements.txt << 'EOF'
pandas==1.5.3
numpy==1.24.3
pydantic==1.10.8
duckdb==0.9.2
pyyaml==6.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
streamlit==1.28.1
plotly==5.17.0
openpyxl==3.1.2
requests==2.31.0
python-dateutil==2.8.2
EOF